backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
blis==1.2.0
boto3==1.35.85
botocore==1.35.85
CacheControl==0.14.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.11
datasets==3.1.0
decorator==5.1.1
defusedxml==0.7.1
dict-toolbox==5.0.0
dill==0.3.8
docopt==0.6.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.14.0
firebase==4.0.1
firebase-admin==6.6.0
flwr==1.15.2
flwr-datasets==0.5.0
fonttools==4.53.0
frozenlist==1.5.0
fsspec==2024.6.0
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.29.1
idna==3.7
ipython==8.12.3
iterators==0.0.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.10.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
# Editable install with no version control (my-app==1.0.0)
-e /Users/vishwajithp/flower_learn/my-app
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.10.0.84
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.3.6
pop-config==12.0.4
pop-loop==1.1.0
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.25.0
protobuf==4.25.6
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pygame==2.5.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.1.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzbar==0.1.9
pyzmq==26.2.0
ray==2.31.0
referencing==0.35.1
rend==7.0.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
thinc==8.3.4
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.28
ultralytics-thop==0.2.7
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.17.2
xxhash==3.5.0
yarg==0.1.9
yarl==1.18.3
fuzzywuzzy
yahooquery
pandas